Use `Cow<'a, str>` instead of `String` where possible (#12)
In the [Vector](https://github.com/vectordotdev/vector) project we've noticed
that user agent parsing results in a lot of small string allocations. This
commit adjusts the API of this crate to avoid allocating small strings where
possible, unfortunately breaking the API contract with existing
users. Performance as measured by the microbenchmarks does appear to be overall
improved, but I will also confirm this in Vector's performance testing harness.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>